Change treatment of multiple Via: addresses
Previously, "Via: [A], [B], [C]" was interpreted as
  Via: [A]
  Via: [B]
  Via: [C]

and similarly the API did not expose a mechanism for adding extra
via-params to an existing header, only new Via headers.

Since it transpires that UACs SHOULD discard messages with multiple
Via headers (RFC 3261 8.1.3.3), we shouldn't treate multiple sections
and multiple headers as interchangeable in this case.

Via headers are now a list of pointers to ViaSections.