Cherry-pick header handling code from master.
Cherry-pick the code from master that moves headers to being stored as
a map, with a slice for order.
Add convenience functions for manipulating the headers.
Make the map and slice themselves private.
Fix up existing code to use the new functions for handling headers.
Put headers in a map so they can be accessed by name

Cherry-picks:
 - Put headers into a map so they can be accessed by name
   (575c850c6012a6e88b5b5f177c6d2b0469caaec0)

 - Added header ordering
   (a6bcd2949c5dd447b9a467125c2bc6731374ccf6)

Conflicts:
	parser/parser.go