Fix mass UDP test
Turns out it wasn't that the receiving routine was too quick to time
out.
In fact, the problem was that sending 10000 messages at once overruns
Linux's packet buffer, and the packets get dropped. Fix is to wait a
millisecond every 100 sends to give the buffer a chance to recover.
Confusion arose because this doesn't happen on Windows.