Make timer calls in transport library mockable
Create new 'timing' library which offers a wrapper around realtime
functions in the built-in 'time' library, e.g. NewTimer, Now, After,
etc. If the 'MockMode' flag in the library is set (at runtime),
subsequent calls to these methods will no longer use clock time; instead
time will only progress when timing.Elapse is called. This allows
precise control of the passage of time in tests.

Modify the existing transport code to call through to 'timing' for these
methods rather than 'time' so that they can be mocked for tests.
Continue to use the 'time' library to refer to constants, types, and
non-realtime functions.