Issue 13 - ConnTable can panic when it is stopped
If at least one connection in the connTable has expired, the
corresponding connWatcher will panic when it is stopped.
This is because the connWatcher nils out its connection both when it
expires, and when the table is stopped - causing a nil pointer
dereference the second time around.

Also, rewrite the entire threading model of connTable because it's full
of race conditions at the moment.